feat: add a method in the MapSystem to now if a cell is empty
